[BUGFIX] moved reading of current sysTick into the GLOBAL_LOCK-Block.
Bug description:
there was a possibility, that during reading, or shortly after, the current sysTick, a thread preempetion or interrupt was generated. during the preemption or interrupt HAL_Time_ExtendedCurrentSysTicks was called again and the values of the kernelTick and the prevKernelTick changed after completion. As the program jumped back to the last thread the data of prevKernelTick was not coherent to the previous call and an overflow was detected, due to the if in line 29. the sysTick was then handled as if 50 days had passed and comming back from the function resultet in all timeouts would trigger, so the socketexception "timedout" was thrown and passed to the c# application.